requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
background==0.2.1
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.0
cycler==0.12.1
daphne==4.0.0
Django==5.0.1
django-background-tasks==1.2.5
django-compat==1.0.15
dwebsocket==0.5.12
Faker==18.10.0
fonttools==4.47.2
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
incremental==22.10.0
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
mysqlclient==2.2.1
nest-asyncio==1.6.0
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyee==8.2.2
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
selenium==4.17.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
websockets==10.4
wordcloud==1.9.3
wsproto==1.2.0
zipp==3.17.0
zope.interface==6.1
